+++ lib/src/extensions/index.dart
@@ -1,10 +1,10 @@
 export 'date_time.dart';
 export 'directory.dart';
 export 'double.dart';
+export 'duration.dart';
 export 'file.dart';
 export 'int.dart';
 export 'iterable.dart';
 export 'list.dart';
 export 'map.dart';
-export 'num.dart';
 export 'string.dart';