+++ lib/vit_dart_extensions.dart
@@ -1,10 +1,14 @@
 library;
 
+// Individual exports are used here over exporting the `index.dart` file to
+// prevent file and directory extensions to be default exported and causing
+// web build to crash.
+
 export 'src/extensions/date_time.dart';
 export 'src/extensions/double.dart';
+export 'src/extensions/duration.dart';
 export 'src/extensions/int.dart';
 export 'src/extensions/iterable.dart';
 export 'src/extensions/list.dart';
 export 'src/extensions/map.dart';
-export 'src/extensions/num.dart';
 export 'src/extensions/string.dart';